nino working on

+++ ninoBot_main.py
@@ -14,20 +14,23 @@ bot = discord.Bot(intents=intents)
 
 @bot.event
 async def on_ready():
-    global client
-    client = Client("https://fffiloni-magnet.hf.space/--replicas/58jap/")
+    global clientMAGNeT, magnetAvailable
+    
+    try:
+        clientMAGNeT = Client("https://fffiloni-magnet.hf.space/--replicas/58jap/")
+        magnetAvailable = True
+    except Exception as e:
+        print(f"Failed to connect to MAGNet: {str(e)}")
+        magnetAvailable = False
+
     print(f'{bot.user} has connected to Discord!')
 
 @bot.command(name="ninotalk", description="Send a message with Nino to a specific channel")
 @commands.has_role(1121063615106142329)
 # pycord will figure out the types for you
 async def add(ctx, channel: discord.TextChannel, message: discord.Option(str)):
-    # you can use them as they were actual integers
-    if str(channel.id) == "1121067425731989524":
-        await channel.send(message)
-        await ctx.respond(f'<:NinoBlue:1205472560842539028> successfuly sent the message "{message}" sent to <#{channel.id}> !')
-    else:
-        await ctx.respond("You can use the `/ninotalk` command only in <#1205476033642496030>", ephemeral=True)
+    await channel.send(message)
+    await ctx.respond(f'<:NinoBlue:1205472560842539028> successfuly sent the message `{message}` sent to <#{channel.id}> !')
 
 @add.error
 async def your_command_error(ctx, error):
@@ -38,50 +41,52 @@ async def your_command_error(ctx, error):
 @bot.command(name="ninoaudio", description="Generate audio tracks from text")
 @commands.check_any(commands.has_role(1121063615106142329), commands.has_role(1204421985443258498))
 async def add(ctx, model: discord.Option(str, "Choose your model", choices=["facebook/magnet-small-10secs", "facebook/magnet-medium-10secs", "facebook/magnet-small-30secs", "facebook/magnet-medium-30secs", "facebook/audio-magnet-small", "facebook/audio-magnet-medium"]), prompt: discord.Option(str)):
-    await ctx.respond(f'<:NinoBlue:1205472560842539028> successfuly added in queue!\nYour audio generation with the prompt `{prompt}` using the model {model} has started, in approx. 45s will be available!', ephemeral=True)
-    result = client.predict(
-		model,	# Literal['facebook/magnet-small-10secs', 'facebook/magnet-medium-10secs', 'facebook/magnet-small-30secs', 'facebook/magnet-medium-30secs', 'facebook/audio-magnet-small', 'facebook/audio-magnet-medium']  in 'Model' Radio component
-		"",	# str  in 'Model Path (custom models)' Textbox component
-		prompt,	# str  in 'Input Text' Textbox component
-		3,	# float  in 'Temperature' Number component
-		0.9,	# float  in 'Top-p' Number component
-		10,	# float  in 'Max CFG coefficient' Number component
-		1,	# float  in 'Min CFG coefficient' Number component
-		20,	# float  in 'Decoding Steps (stage 1)' Number component
-		10,	# float  in 'Decoding Steps (stage 2)' Number component
-		10,	# float  in 'Decoding Steps (stage 3)' Number component
-		10,	# float  in 'Decoding Steps (stage 4)' Number component
-		"prod-stride1 (new!)",	# Literal['max-nonoverlap', 'prod-stride1 (new!)']  in 'Span Scoring' Radio component
-		api_name="/predict_full"
-    )
+    if magnetAvailable:
+        await ctx.respond(f'<:NinoBlue:1205472560842539028> successfuly added in queue!\nYour audio generation with the prompt `{prompt}` using the model `{model}` has started, in approx. 45s will be available!', ephemeral=True)
+        result = clientMAGNeT.predict(
+            model,	# Literal['facebook/magnet-small-10secs', 'facebook/magnet-medium-10secs', 'facebook/magnet-small-30secs', 'facebook/magnet-medium-30secs', 'facebook/audio-magnet-small', 'facebook/audio-magnet-medium']  in 'Model' Radio component
+            "",	# str  in 'Model Path (custom models)' Textbox component
+            prompt,	# str  in 'Input Text' Textbox component
+            3,	# float  in 'Temperature' Number component
+            0.9,	# float  in 'Top-p' Number component
+            10,	# float  in 'Max CFG coefficient' Number component
+            1,	# float  in 'Min CFG coefficient' Number component
+            20,	# float  in 'Decoding Steps (stage 1)' Number component
+            10,	# float  in 'Decoding Steps (stage 2)' Number component
+            10,	# float  in 'Decoding Steps (stage 3)' Number component
+            10,	# float  in 'Decoding Steps (stage 4)' Number component
+            "prod-stride1 (new!)",	# Literal['max-nonoverlap', 'prod-stride1 (new!)']  in 'Span Scoring' Radio component
+            api_name="/predict_full"
+        )
 
-    print("Result generated!")
+        print(f"{prompt} from {ctx.author} generated!")
 
-    # Format the result paths
-    parent_folders = []
-    result_paths = []
-    if isinstance(result[0], dict):
-        for key, value in result[0].items():
-            if key != "subtitles":
-                os.remove(value)
-                os.rmdir(os.path.dirname(value))
-    for path in result[1:]:
-        result_paths.append(path)
+        # Format the result paths
+        parent_folders = []
+        result_paths = []
+        if isinstance(result[0], dict):
+            for key, value in result[0].items():
+                if key != "subtitles":
+                    os.remove(value)
+                    os.rmdir(os.path.dirname(value))
+        for path in result[1:]:
+            result_paths.append(path)
 
-    try:
-        await ctx.respond(f"<:NinoBlue:1205472560842539028> `{prompt}` generated using the MAGNeT model `{model}` by <@{ctx.author.id}>")
-        for index, path in enumerate(result_paths):
-            with open(path, 'rb') as fp:
-                await ctx.followup.send(file=File(fp, f"{prompt}_{ctx.author}_output_{index}.wav"))
-                os.remove(path)
-                parent_folders.append(os.path.dirname(path))
-        
-        for folder in parent_folders:
-            os.rmdir(folder)
-        
-    except Exception as e:
-        await ctx.respond(f"Failed to send the audio file: {str(e)}", ephemeral=True)
-
+        try:
+            await ctx.respond(f"<:NinoBlue:1205472560842539028> `{prompt}` generated using the MAGNeT model `{model}` by <@{ctx.author.id}>")
+            for index, path in enumerate(result_paths):
+                with open(path, 'rb') as fp:
+                    await ctx.followup.send(file=File(fp, f"{prompt}_{ctx.author}_output_{index}.wav"))
+                    os.remove(path)
+                    parent_folders.append(os.path.dirname(path))
+            
+            for folder in parent_folders:
+                os.rmdir(folder)
+            
+        except Exception as e:
+            await ctx.respond(f"Failed to send the audio file: {str(e)}", ephemeral=True)
+    else:
+        await ctx.respond("The MAGNeT service is currently unavailable, please try again later.", ephemeral=True)
     
 
 @add.error
